refactor: 도면 이미지 유틸 역할 명확화
